requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
datasets==2.19.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==6.0.0
elastic-transport==8.4.1
elasticsearch==8.6.1
email_validator==2.1.1
esrally==2.10.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==1.22.1
google-crc32c==1.5.0
google-resumable-media==1.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
ijson==2.6.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==2.11.3
jsonschema==3.1.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.7.5
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
numpy==1.24.4
openai==0.28.0
orjson==3.10.3
packaging==24.0
pandas==2.0.3
pillow==10.3.0
psutil==5.9.4
py-cpuinfo==7.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tabulate==0.8.9
thespian==3.10.1
tk==0.1.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
xxhash==3.4.1
yappi==1.4.0
yarl==1.9.4
zipp==3.18.2
zstandard==0.21.0
